@@ -1,5 +1,11 @@
 package gmicro
 
+import (
+	"strings"
+)
+
+const newlineChar = "\n"
+
 // Logger is logger interface.
 type Logger interface {
 	Printf(string, ...interface{})
@@ -9,7 +15,13 @@ type Logger interface {
 type LoggerFunc func(string, ...interface{})
 
 // Printf implements Logger interface.
-func (f LoggerFunc) Printf(msg string, args ...interface{}) { f(msg, args...) }
+func (f LoggerFunc) Printf(msg string, args ...interface{}) {
+	if !strings.HasSuffix(msg, newlineChar) {
+		msg += newlineChar
+	}
+
+	f(msg, args...)
+}
 
 // dummy logger writes nothing.
 var dummyLogger = LoggerFunc(func(string, ...interface{}) {})
