@@ -1,26 +1,18 @@
 package gmicro
 
 import (
-	"strings"
-
-	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"github.com/grpc-ecosystem/grpc-gateway/utilities"
+	gRuntime "github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
+	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
 )
 
 // Route represents the route for mux
 type Route struct {
 	Method  string
-	Pattern runtime.Pattern
-	Handler runtime.HandlerFunc
-}
-
-// PathPattern returns a pattern which matches exactly with the path
-func PathPattern(path string) runtime.Pattern {
-	path = strings.TrimPrefix(path, "/")
-	return runtime.MustPattern(runtime.NewPattern(1, []int{int(utilities.OpLitPush), 0}, []string{path}, ""))
+	Path    string
+	Handler gRuntime.HandlerFunc
 }
 
 // AllPattern returns a pattern which matches any url
-func AllPattern() runtime.Pattern {
-	return runtime.MustPattern(runtime.NewPattern(1, []int{int(utilities.OpPush), 0}, []string{""}, ""))
+func AllPattern() gRuntime.Pattern {
+	return gRuntime.MustPattern(gRuntime.NewPattern(1, []int{int(utilities.OpPush), 0}, []string{""}, ""))
 }
